+++ controllers/index.js
@@ -1,5 +1,13 @@
-const contacts = require("./contacts");
+const getAll = require("./getAll");
+const getById = require("./getById");
+const add = require("./add");
+const updateById = require("./updateById");
+const deleteById = require("./deleteById");
 
 module.exports = {
-  contacts,
+  getAll,
+  getById,
+  add,
+  updateById,
+  deleteById,
 };
