+++ model/contacts/index.js
@@ -1,0 +1,13 @@
+const getListContacts = require("./getListContacts");
+const getContactById = require("./getContactById");
+const addContact = require("./addContact");
+const updateContact = require("./updateContact");
+const removeContact = require("./removeContact");
+
+module.exports = {
+  getListContacts,
+  getContactById,
+  addContact,
+  updateContact,
+  removeContact,
+};
